+++ langs/fire/fire.test.js
@@ -1,12 +1,12 @@
 #! /usr/local/bin/node --use_strict
 
-const FireProgram = require("./FireProgram.js")
+const fire = require("./fire.js")
 
 const testTree = {}
 
 testTree.all = equal => {
   // Arrange
-  const program = new FireProgram(`add ten 2 3 5`)
+  const program = new fire(`add ten 2 3 5`)
 
   // Act/Assert
   equal(program.getAllErrors().length, 0, "no errors")