+++ langs/numbers/numbers.test.js
@@ -1,12 +1,12 @@
 #! /usr/local/bin/node --use_strict
 
-const NumbersProgram = require("./Numbers.js").Program
+const numbers = require("./numbers.js")
 
 const testTree = {}
 
 testTree.all = equal => {
   // Arrange
-  const program = new NumbersProgram(`+ 2 2 1 2
+  const program = new numbers(`+ 2 2 1 2
 - 2 2`)
 
   // Act/Assert
@@ -14,7 +14,7 @@ testTree.all = equal => {
   equal(program.executeSync().join(" "), `7 0`)
 
   // A/A/A
-  equal(new NumbersProgram(`+ 2 2 1 1`).executeSync().join(""), `6`)
+  equal(new numbers(`+ 2 2 1 1`).executeSync().join(""), `6`)
 }
 
 /*NODE_JS_ONLY*/ if (!module.parent) require("../../tests/testTreeRunner.js")(testTree)